Rename code files in Ui/Windows/
