+++ src/OpenLoco/CMakeLists.txt
@@ -315,9 +315,9 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/AboutMusic.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/BuildVehicle.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Cheats.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Company.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyFaceSelection.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyList.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyWindow.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Common.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/ConstructionTab.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/OverheadTab.cpp"
@@ -326,35 +326,36 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/DebugWindow.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/DragVehiclePart.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/EditKeyboardShortcut.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/EditorStepNavigation.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Error.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Industry.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/IndustryList.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/IndustryWindow.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/KeyboardShortcuts.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/LandscapeGeneration.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/LandscapeGenerationConfirm.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Main.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/MapToolTip.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/MapWindow.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/MessageWindow.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Map.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/MapTooltip.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Messages.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/MusicSelection.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/NetworkStatus.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/News/Common.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/News/News.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/News/Ticker.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ObjectLoadError.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ObjectSelectionWindow.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ObjectSelection.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Options.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PlayerInfoPanel.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ProgressBar.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptBrowseWindow.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptOkCancelWindow.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptSaveWindow.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptBrowse.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptOkCancel.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/PromptSave.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ScenarioOptions.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ScenarioSelect.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Station.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/StationList.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/StationWindow.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TerraForm.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TextInputWindow.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Terraform.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TextInput.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TileInspector.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TimePanel.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TitleExit.cpp"
@@ -362,13 +363,12 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TitleMenu.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TitleOptions.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TitleVersion.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolTip.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarBottomEditor.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarTop.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarTopAlt.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarTopCommon.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Toolbar/Common.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Toolbar/Editor.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Toolbar/Game.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Tooltip.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Town.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TownList.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/TownWindow.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Tutorial.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Vehicle.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/VehicleList.cpp"
@@ -703,7 +703,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowType.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Construction.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/News/News.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarTopCommon.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Toolbar/Common.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/OrderManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Orders.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Routing.h"
