+++ src/OpenLoco/src/Ui/Windows/EditorStepNavigation.cpp
