+++ src/OpenLoco/src/Ui/Windows/Industry.cpp
