+++ src/OpenLoco/src/Ui/Windows/Map.cpp
