+++ src/OpenLoco/src/Ui/Windows/ObjectSelection.cpp
