+++ src/OpenLoco/src/Ui/Windows/Toolbar/Common.cpp
@@ -1,4 +1,4 @@
-#include "ToolbarTopCommon.h"
+#include "Common.h"
 #include "Audio/Audio.h"
 #include "Config.h"
 #include "Entities/EntityManager.h"
