+++ src/OpenLoco/src/Ui/Windows/Toolbar/Common.h
