+++ src/OpenLoco/src/Ui/Windows/Toolbar/Game.cpp
@@ -22,7 +22,7 @@
 #include "Objects/WaterObject.h"
 #include "S5/S5.h"
 #include "SceneManager.h"
-#include "ToolbarTopCommon.h"
+#include "Common.h"
 #include "Ui/Dropdown.h"
 #include "Ui/Screenshot.h"
 #include "Ui/ToolManager.h"
